# frontend/api/openapi.yml
openapi: 3.0.0
info:
  title: SyncFam API
  version: 1.0.0
servers:
  - url: /api/v1
    description: Development server

components:
  schemas:
    # --- Common Wrapper Schema ---
    SuccessResponse:
      type: object
      properties:
        data:
          description: APIコールが成功した場合の実際のデータ
          # 型は各APIのレスポンス定義で上書き or anyOf/oneOf などで指定
      required:
        - data
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: エラーメッセージ
          example: '入力内容に誤りがあります。'
        details:
          # バリデーションエラー等の詳細を格納 (キーがフィールド名、値がエラーメッセージなど)
          type: object
          additionalProperties:
            type: string # または array of string
          example:
            {
              'name': 'タスク名は必須です。',
              'dueDate': '日付の形式が正しくありません。'
            }
      required:
        - message

    # --- ページネーション情報を含むレスポンススキーマ ---
    PaginationMeta:
      type: object
      properties:
        totalItems:
          type: integer
          description: 条件に一致する総アイテム数
          example: 120
        totalPages:
          type: integer
          description: 総ページ数
          example: 12
        currentPage:
          type: integer
          description: 現在のページ番号 (1始まり)
          example: 1
        limit:
          type: integer
          description: 1ページあたりのアイテム数
          example: 10
      required:
        - totalItems
        - totalPages
        - currentPage
        - limit
    PaginatedTasksResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Task'
          description: 現在のページのタスクデータ
        meta:
          $ref: '#/components/schemas/PaginationMeta' # ページネーション情報
      required:
        - data
        - meta

    # --- Label Schema ★★★ ---
    Label:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: ラベルの一意なID
          readOnly: true
        name:
          type: string
          description: ラベル名
          example: "家事"
        color: # (任意)
          type: string
          description: "ラベルの色 (例: HEXカラーコード)"
          example: "#4A90E2"
          nullable: true
        # createdAt, updatedAt は API レスポンスに含めなくても良いかも
      required:
        - id
        - name

    # --- Task Schemas ---
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: タスクの一意なID (サーバーで自動採番)
          example: 'a1b2c3d4-e5f6-7890-1234-567890abcdef'
          readOnly: true # クライアントからは送信されず、サーバーが生成することを示す
        name:
          type: string
          description: タスク名
          example: '買い物に行く'
        assigneeId:
          type: string
          format: uuid
          nullable: true
          description: 担当者のユーザーID (nullの場合は未割り当て)
          example: 'f0e9d8c7-b6a5-4321-fedc-ba9876543210'
        dueDate:
          type: string
          format: date
          nullable: true
          description: 期限日 (YYYY-MM-DD)
          example: '2025-04-30'
        isCompleted:
          type: boolean
          description: 完了状態フラグ
          example: false
        labels:
          type: array
          items:
            $ref: '#/components/schemas/Label' # Label スキーマを参照
          description: タスクに付与されたラベルのリスト
        isRecurring:
          type: boolean
          description: このタスクが定常タスクかどうか
          default: false
        recurrenceRule:
          type: string
          description: '繰り返しのルール'
          nullable: true
          example: 'daily'
        createdAt:
          type: string
          format: date-time
          description: 作成日時
          example: '2025-04-20T10:00:00Z'
          readOnly: true # 作成日時は通常サーバーで設定
        updatedAt:
          type: string
          format: date-time
          description: 更新日時
          example: '2025-04-20T11:00:00Z'
          readOnly: true # 更新日時は通常サーバーで設定
      required:
        - id
        - name
        - isCompleted
        - labels
        - createdAt
        - updatedAt

    # Task操作
    CreateTaskInput:
      type: object
      properties:
        name:
          type: string
          description: タスク名 (必須)
          example: '新しいタスク'
          minLength: 1 # バリデーション例
          maxLength: 100
        assigneeId:
          type: string
          format: uuid
          nullable: true
          description: 担当者のユーザーID (null許容)
        dueDate:
          type: string
          format: date
          nullable: true
          description: 期限日 (YYYY-MM-DD, null許容)
          # pattern: '^\d{4}-\d{2}-\d{2}$' # 正規表現でのバリデーション例
        label_ids:
          type: array
          items:
            type: string
            format: uuid
          description: 紐付けるラベルのID配列 (必須、最低1つ)
          minItems: 1 # 最低1つのラベルIDを必須とする
          example: ["label-uuid-1", "label-uuid-2"]
      required:
        - name
        - label_ids
    PutTaskInput:
      type: object
      description: 更新後のタスクの完全な状態 (id, createdAt, updatedAtを除く)
      properties:
        name:
          type: string
          description: タスク名 (必須)
          example: '更新されたタスク名'
          minLength: 1
          maxLength: 100
        assigneeId:
          type: string
          format: uuid
          nullable: true
          description: 担当者のユーザーID (null許容)
        dueDate:
          type: string
          format: date
          nullable: true
          description: 期限日 (YYYY-MM-DD, null許容)
        isCompleted:
          type: boolean
          description: 完了状態フラグ (必須)
          example: true
        label_ids:
          type: array
          items:
            type: string
            format: uuid
          description: 更新後のラベルID配列 (これで既存の紐付けは上書きされる。必須、最低1つ)
          minItems: 1 # 最低1つのラベルIDを必須とする
          example: ["label-uuid-3"]
      required:
        - name
        - isCompleted
        - label_ids

    # User操作
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: ユーザーの一意なID
          example: 'f0e9d8c7-b6a5-4321-fedc-ba9876543210'
          readOnly: true
        name:
          type: string
          description: ユーザー名 (表示名)
          example: '田中 太郎'
        # 必要に応じて他のユーザー情報を追加 (例: email, avatarUrl)
        # email:
        #   type: string
        #   format: email
        # avatarUrl:
        #   type: string
        #   format: url
        #   nullable: true
        createdAt:
          type: string
          format: date-time
          description: アカウント作成日時
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          description: アカウント更新日時
          readOnly: true
      required:
        - id
        - name
        - createdAt
        - updatedAt

    # Label 操作
    LabelCreateInput:
      type: object
      properties:
        name:
          type: string
          description: 新しいラベル名 (必須)
          minLength: 1
          maxLength: 50 # 例
        color:
          type: string
          description: ラベルの色 (任意)
          nullable: true
      required:
        - name
    LabelUpdateInput:
      type: object
      properties:
        name:
          type: string
          description: 更新後のラベル名 (任意)
          minLength: 1
          maxLength: 50
        color:
          type: string
          description: 更新後のラベルの色 (任意)
          nullable: true
      minProperties: 1 # 最低1つのプロパティが必要

  parameters:
    TaskIdParam:
      name: taskId
      in: path
      required: true
      description: 操作対象のタスクID
      schema:
        type: string
        format: uuid
    PaginationPageParam:
      name: page
      in: query
      required: false
      description: 表示するページ番号 (1始まり)
      schema:
        type: integer
        minimum: 1
        default: 1
    PaginationLimitParam:
      name: limit
      in: query
      required: false
      description: 1ページあたりのアイテム数
      schema:
        type: integer
        minimum: 1
        maximum: 100 # 上限を設定
        default: 20 # デフォルト値を設定

    # ★★★ GET /tasks 用のパラメータ定義 ★★★
    TaskAssigneeIdFilterParam:
      name: assigneeId
      in: query
      required: false
      description: 担当者IDでフィルタリング ('me' で自分のタスクを指定可能)
      schema:
        type: string
        example: 'me'
    TaskIsCompletedFilterParam:
      name: isCompleted
      in: query
      required: false
      description: 完了状態でフィルタリング (true/false)
      schema:
        type: boolean
    TaskSortParam:
      name: sort
      in: query
      required: false
      description: ソート順を指定
      schema:
        type: string
        enum:
          - createdAt_desc
          - createdAt_asc
          - dueDate_asc
          - dueDate_desc
        default: createdAt_desc
    TaskLabelsFilterParam:
      name: labels
      in: query
      required: false
      description: 指定されたラベルを**すべて**含むタスクをフィルタリング (カンマ区切り)
      schema:
        type: string # カンマ区切りの文字列として渡す
        example: "家事,重要"

    LabelIdParam:
      name: labelId
      in: path
      required: true
      description: 操作対象のラベルID
      schema:
        type: string
        format: uuid

paths:
  /health:
    get:
      summary: Health Check
      tags: [Meta]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object # SuccessResponse でラップしない例 (ヘルスチェック等)
                properties:
                  status:
                    type: string
                    example: OK

  # --- Tasks API ---
  /tasks:
    get:
      summary: タスク一覧取得
      tags: [Tasks]
      parameters:
       - $ref: '#/components/parameters/TaskAssigneeIdFilterParam'
       - $ref: '#/components/parameters/TaskIsCompletedFilterParam'
       - $ref: '#/components/parameters/TaskSortParam'
       - $ref: '#/components/parameters/TaskLabelsFilterParam'
       - $ref: '#/components/parameters/PaginationPageParam'
       - $ref: '#/components/parameters/PaginationLimitParam'
      responses:
        '200':
          description: タスク一覧取得成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedTasksResponse'
        '400': # パラメータ不正などの場合
          description: Bad Request (Invalid query parameters)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: 新規タスク作成
      tags: [Tasks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskInput'
      responses:
        '201':
          description: タスク作成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Bad Request (バリデーションエラーなど)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /tasks/{taskId}:
    parameters:
      - $ref: '#/components/parameters/TaskIdParam'
    put:
      summary: タスク更新 (全量)
      tags: [Tasks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PutTaskInput'
      responses:
        '200': # 成功時は 200 OK で更新後のリソースを返す
          description: タスク更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Bad Request (バリデーションエラーなど)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: タスク削除
      tags: [Tasks]
      responses:
        '204':
          description: タスク削除成功 (No Content) # ボディは返さない
        '404':
          description: Task Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # --- Users API ---
  /members:
    get:
      summary: 家族メンバー一覧取得
      tags: [Members]
      description: 現在認証されているユーザーが所属する家族（または共有グループ）のメンバー一覧を取得します。担当者選択などで使用します。
      # parameters: # 必要に応じて検索やページネーションパラメータを追加
      #   - name: search
      #     in: query
      #     schema:
      #       type: string
      responses:
        '200':
          description: メンバー一覧取得成功
          content:
            application/json:
              schema:
                # SuccessResponseでラップし、dataの中身を定義
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/User' # Userスキーマの配列
                    required:
                      - data
              # レスポンスサンプル (任意)
              examples:
                default:
                  value:
                    data:
                      - id: 'f0e9d8c7-b6a5-4321-fedc-ba9876543210'
                        name: '田中 太郎'
                        createdAt: '2025-01-10T10:00:00Z'
                        updatedAt: '2025-04-15T11:00:00Z'
                      - id: 'e1d2c3b4-a598-7654-3210-fedcba98765f'
                        name: '山田 花子'
                        createdAt: '2025-02-20T15:30:00Z'
                        updatedAt: '2025-03-01T09:45:00Z'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # --- Labels API ---
  /labels:
    get:
      summary: ラベル一覧取得
      tags: [Labels]
      description: 利用可能なすべてのラベルを取得します。
      responses:
        '200':
          description: ラベル一覧取得成功
          content:
            application/json:
              schema:
                type: object # SuccessResponse でラップしても良い
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Label'
                required:
                  - data
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: 新規ラベル作成
      tags: [Labels]
      description: 新しいラベルを作成します。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LabelCreateInput'
      responses:
        '201':
          description: ラベル作成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Label' # 作成されたラベルを返す
        '400': # バリデーションエラーや名前重複など
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'